+++ src/test/java/com/epam/reportportal/junit/ItemTimeOrderTest.java
@@ -60,7 +60,7 @@ public class ItemTimeOrderTest {
 	public void setupMock() {
 		TestUtils.mockLaunch(client, null, firstSuiteId, secondSuiteId, classIds);
 		TestUtils.mockNestedSteps(client, tests);
-		TestUtils.mockLogging(client);
+		TestUtils.mockBatchLogging(client);
 		ReportPortalListener.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters()));
 	}