+++ src/test/java/com/epam/reportportal/junit/nested/NestedStepPassedTest.java
@@ -43,7 +43,7 @@ public class NestedStepPassedTest {
 	@BeforeEach
 	public void setupMock() {
 		TestUtils.mockLaunch(client, null, classId, methodId, nestedId);
-		TestUtils.mockLogging(client);
+		TestUtils.mockBatchLogging(client);
 		ReportPortalListener.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters()));
 	}