+++ src/test/java/com/epam/reportportal/junit/testcaseid/TestCaseIdStaticJunitParamsTest.java
@@ -49,7 +49,7 @@ public class TestCaseIdStaticJunitParamsTest {
 	public void setupMock() {
 		client = mock(ReportPortalClient.class);
 		TestUtils.mockLaunch(client, null, null, classId, methodId);
-		TestUtils.mockLogging(client);
+		TestUtils.mockBatchLogging(client);
 		ReportPortalListener.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters()));
 	}