+++ src/test/java/com/epam/reportportal/junit/testcaseid/TestCaseIdStaticTest.java
@@ -43,7 +43,7 @@ public class TestCaseIdStaticTest {
 	public void setupMock() {
 		client = mock(ReportPortalClient.class);
 		TestUtils.mockLaunch(client, null, null, classId, methodId);
-		TestUtils.mockLogging(client);
+		TestUtils.mockBatchLogging(client);
 		ReportPortalListener.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters()));
 	}